refactor: standardize song ID to Long
Consistency across data access methods by replacing Int-based ID references with Long to match the database schema.

+++ app/src/main/java/io/github/yisus/avenor/ui/screens/PlaylistDetailsScreen.kt
@@ -136,7 +136,7 @@ fun PlaylistDetailsScreen(viewModel: PlaybackViewModel, playlistId: Int) {
                                     initialName = song.title,
                                     onDismiss = { showRename = false },
                                     onRename = { newName: String -> 
-                                        viewModel.renameSong(song.id.toInt(), newName)
+                                        viewModel.renameSong(song.id, newName)
                                         showRename = false
                                     }
                                 )
